refactor: add repo service and update related code
