+++ src/renderer/plugins/markdown-link.ts
@@ -7,7 +7,8 @@ import { isElectron, isWindows } from '@fe/support/env'
 import { useToast } from '@fe/support/ui/toast'
 import { DOM_ATTR_NAME, DOM_CLASS_NAME } from '@fe/support/args'
 import { basename, dirname, join, normalizeSep, resolve } from '@fe/utils/path'
-import { getAttachmentURL, getRepo, openExternal, openPath } from '@fe/services/base'
+import { getAttachmentURL, openExternal, openPath } from '@fe/services/base'
+import { getRepo } from '@fe/services/repo'
 import { getAllCustomEditors } from '@fe/services/editor'
 import { fetchTree } from '@fe/support/api'
 import type { Doc } from '@share/types'
