+++ src/renderer/services/document.ts
@@ -19,7 +19,8 @@ import * as api from '@fe/support/api'
 import { fileToBase64URL, getLogger } from '@fe/utils'
 import { isWindows } from '@fe/support/env'
 import CreateFilePanel from '@fe/components/CreateFilePanel.vue'
-import { getAllRepos, getRepo, inputPassword, openPath, showItemInFolder } from './base'
+import { inputPassword, openPath, showItemInFolder } from './base'
+import { getAllRepos, getRepo } from './repo'
 import { $$t, t } from './i18n'
 import { getSetting, setSetting } from './setting'
